move components into relevant places
