+++ __tests__/svg_art.t.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { render } from "@testing-library/react";
 
-import ShareCassette from "@/components/ShareCassette";
-import ShareUnify from "@/components/ShareUnify";
+import ShareCassette from "@/app/user/[slug]/ShareCassette";
+import ShareUnify from "@/app/unify/[users]/ShareUnify";
 
 describe("ShareCassette", () => {
   it("should import SVG content as a module", () => {
