+++ src/app/page.jsx
@@ -7,10 +7,10 @@ Home/Index page of application, contains buttons to log in in with Spotify and s
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import loginWithSpotify from "@/app/login/actions";
-import Ipod from "@/components/IPod";
+import Ipod from "@/app/IPod";
 import "@/app/globals.css";
-import LeftPanel from "@/components/LeftPanel";
-import LoadingIcon from "@/components/LoadingIcon";
+import LeftPanel from "@/app/LeftPanel";
+import LoadingIcon from "@/app/LoadingIcon";
 import createClient from "@/utils/supabase/client";
 
 export default function IndexPage() {
